Fix division by 0 when a steppable action has no items to migrate
Introduce AbstractSteppable class not to repeat ourselves
